fix (dev): rename directory to temporary name for fix
