+++ point-system/src/main/java/pointsystem/converter_temp/CompanyConverter.java
@@ -1,4 +1,4 @@
-package pointsystem.converter;
+package pointsystem.converter_temp;
 
 import org.modelmapper.ModelMapper;
 import org.modelmapper.TypeMap;
