+++ point-system/src/main/java/pointsystem/converter_temp/Converter.java
@@ -1,4 +1,4 @@
-package pointsystem.converter;
+package pointsystem.converter_temp;
 
 import java.util.List;
