+++ point-system/src/main/java/pointsystem/converter_temp/UserConverter.java
@@ -1,4 +1,4 @@
-package pointsystem.converter;
+package pointsystem.converter_temp;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
