+++ point-system/src/main/java/pointsystem/service/EmployeeService.java
@@ -5,7 +5,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
-import pointsystem.converter.EmployeeConverter;
+import pointsystem.converter_temp.EmployeeConverter;
 import pointsystem.dto.employee.EmployeeDto;
 import pointsystem.entity.*;
 import pointsystem.repository.CompanyPositionEmployeeRepository;
